Preserve escape sequences
Previously, the string "\n" would be parsed as "n" and printed as "n".
Now, the string "\n" is parsed as "\\n" and is printed as "\n".

The more complete approach would be to parse "\n" as "\n", but this
would involve implementing many different kinds of escape sequences,
such as "\u2e18", and we don't need that kind of accuracy yet.