Patterns, a combinator library
I'm really proud of this one! It's a combinator library for describing
patterns, in the sense of pattern-matching.

Since the implementation is based on my invertible version of lookup,
this means that pattern-matching is invertible! The inverse of
pattern-matching is substitution, in the sense that you can provide
values for each pattern variable and thereby build a closed term.

Using defineIsomorphisms, you can easily create patterns for arbitrary
custom datatypes. One major limitation of the library, however, is that
the holes must all have the same type, i.e. if you have a heterogeneous
pair (1, "a"), you can use the pattern (var "x" <*> pure "a") to match
the 1, you can use the pattern (pure 1 <*> var "y") to match the "a",
but unfortunately you can't use the pattern (var "x" <*> var "y") to
match both.