Extractors, another combinator library
Almost none of the code from the Patterns library was using the fact
that it was an environment which was being carried along. This usually
means that the types overconstrain the implementation, and that there is
something more general hidden under the surface. And indeed there was!
Extractors are more general than Patterns, and also simpler.